Add some basic routes with some simple responses.
Added a get_all style method which returns all servers listed
Added a search_server method which takes a region and gametype and gets all servers in that query.
Added a add_server method which adds a new server given a region, gametype, name, and an IP.

Use postgres directly instead of an ORM like diesel for now.