Diesel initial Migration, routes to a broken out module.
Switching to Diesel will be a lot of good, it seems to be getting better and better, and the code is much better than raw SQL. It enforces strong typing and good SQL practices like migrations. It will also make things like foreign keys and more complex queries easier to manage.

The routes.rs file was getting really large, and it seems that splitting each function to its own file works out really well, so here we are with that.

Still have to do the search_servers functionality, its just a stub right now, couldnt finish it in time for the commit. Wanted at least something committed so here we are. Next commit will have search_servers finished with Diesel.