Further compartmentalize routes into /server and add /region with basic functionality.
The src/routes/ folder was getting a bit ugly, so i decided to put all routes belonging to server functionality into src/routes/server, which will future proof it for other operations. Also, restructuring the source folders brings them in line with the URL path scheme I have going.

The current region functionality is closely related to the equivalent server functionality. Added /region/all to display all regions, and /region/add so I will not have to go into psql CLI to add new regions when I need to. This should be good for now, although I may add a /delete for both servers and regions.

Please forgive me for the tags stuff in update_server, it is really bad, but it works. Need to come up with a cleaner solution than that. Perhaps Serde will save me.